Внес правки согласно замечаниям
